# internal/docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Panahon API.",
        "title": "Panahon API",
        "contact": {
            "name": "Emilio Gozo",
            "email": "emiliogozo@proton.me"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/tokens/renew": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Renew access token",
                "parameters": [
                    {
                        "description": "Renew access token parameters",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RenewAccessTokenParams"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "RenewAccessTokenParams": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}